19.12.2022

+++ RestaurantAPI/Services/RestaurantService.cs
@@ -1,8 +1,12 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
+using RestaurantAPI.Authorization;
 using RestaurantAPI.Entities;
 using RestaurantAPI.Exceptions;
 using RestaurantAPI.Models;
+using System.Linq.Expressions;
+using System.Security.Claims;
 
 namespace RestaurantAPI.Services
 {
@@ -11,7 +15,7 @@ namespace RestaurantAPI.Services
         void Update(int id, UpdateRestaurantDto dto);
         void Delete(int id);
         int Create(CreateRestaurantDto dto);
-        IEnumerable<RestaurantDto> GetAll();
+        PageResult<RestaurantDto> GetAll(RestaurantQuery query);
         RestaurantDto GetById(int id);
     }
 
@@ -20,12 +24,17 @@ namespace RestaurantAPI.Services
         private readonly RestaurantDbContext _context;
         private readonly IMapper _mapper;
         private readonly ILogger<RestaurantService> _logger;
+        private readonly IAuthorizationService _authorizationService;
+        private readonly IUserContextService _userContextService;
 
-        public RestaurantService(RestaurantDbContext context, IMapper mapper, ILogger<RestaurantService> logger)
+        public RestaurantService(RestaurantDbContext context, IMapper mapper, ILogger<RestaurantService> logger,
+            IAuthorizationService authorizationService, IUserContextService userContextService)
         {
             _context = context;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
+            _userContextService = userContextService;
         }
         public void Update(int id, UpdateRestaurantDto dto)
         {
@@ -36,9 +45,16 @@ namespace RestaurantAPI.Services
             if (restaurant is null)
                 throw new NotFoundException("Restaurant not found");
 
-            restaurant.Name= dto.Name;
+            var authorizationResult = _authorizationService.AuthorizeAsync(_userContextService.User, restaurant,
+                new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
+
+            restaurant.Name = dto.Name;
             restaurant.Description = dto.Description;
-            restaurant.HasDelivery =dto.HasDelivery;
+            restaurant.HasDelivery = dto.HasDelivery;
 
             _context.SaveChanges();
         }
@@ -52,6 +68,13 @@ namespace RestaurantAPI.Services
 
             if (restaurant is null)
                 throw new NotFoundException("Restaurant not found");
+
+            var authorizationResult = _authorizationService.AuthorizeAsync(_userContextService.User, restaurant,
+                new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
 
             _context.Restaurants.Remove(restaurant);
             _context.SaveChanges();
@@ -71,21 +94,48 @@ namespace RestaurantAPI.Services
 
             return restaurantDto;
         }
-        public IEnumerable<RestaurantDto> GetAll()
+        public PageResult<RestaurantDto> GetAll(RestaurantQuery query)
         {
-            var restaurants = _context
+            var baseQuery = _context
                 .Restaurants
                 .Include(r => r.Address)
                 .Include(r => r.Dishes)
+                .Where(r => query.SearchPhrase == null || (r.Name.ToLower().Contains(query.SearchPhrase.ToLower())
+                                                    || r.Description.ToLower().Contains(query.SearchPhrase.ToLower())));
+
+            if (!string.IsNullOrEmpty(query.SortBy))
+            {
+                var columnSelector = new Dictionary<string, Expression<Func<Restaurant, object>>>
+                {
+                    { nameof(Restaurant.Name), r => r.Name },
+                    { nameof(Restaurant.Description), r => r.Description },
+                    { nameof(Restaurant.Category), r => r.Category },
+                };
+
+                var selectedColumn = columnSelector[query.SortBy];
+
+                baseQuery = query.SortDirection == SortDirection.ASC
+                    ? baseQuery.OrderBy(selectedColumn)
+                    : baseQuery.OrderByDescending(selectedColumn);
+            }
+
+            var restaurants = baseQuery
+                .Skip(query.PageSize * (query.PageNumber - 1))
+                .Take(query.PageSize)
                 .ToList();
 
-            var restaurantsDto = _mapper.Map<IEnumerable<RestaurantDto>>(restaurants);
+            var totalItemsCount = baseQuery.Count();
 
-            return restaurantsDto;
+            var restaurantsDto = _mapper.Map<List<RestaurantDto>>(restaurants);
+
+            var result = new PageResult<RestaurantDto>(restaurantsDto, totalItemsCount, query.PageSize, query.PageNumber);
+
+            return result;
         }
         public int Create(CreateRestaurantDto dto)
         {
             var restaurant = _mapper.Map<Restaurant>(dto);
+            restaurant.CreatedById = _userContextService.GetUserId;
             _context.Add(restaurant);
             _context.SaveChanges();
